Task/779167000001151013

See merge request eisenvault-dms-main/eisen-syncer!28

+++ desktop/server/helpers/syncers/ondemand.js
@@ -32,9 +32,7 @@ exports.recursiveDownload = async params => {
   logger.info("download step 1");
 
   if (
-    account.sync_enabled == 0 ||
-    (account.sync_in_progress == 1 && recursive === false)
-  ) {
+    account.sync_enabled == 0 || (recursive === false && account.download_in_progress == 1)) {
     logger.info("download bailed");
     return;
   }
@@ -52,7 +50,7 @@ exports.recursiveDownload = async params => {
     return;
   }
 
-  // If the current folder doesnot exists, we will create it.
+  // If the current folder does not exists, we will create it.
   if (!fs.existsSync(`${destinationPath}/${watcher.site_name}/documentLibrary`)) {
     mkdirp(`${destinationPath}/${watcher.site_name}/documentLibrary`);
   }
@@ -71,15 +69,17 @@ exports.recursiveDownload = async params => {
 
     // Check if the node is present in the database
     let record = await nodeModel.getOneByNodeId({
-      account: account,
+      account,
       nodeId: node.id
     });
     logger.info("download step 6");
 
-    if (record && record.upload_in_progress === 1) {
+    if (record && (record.download_in_progress === 1 || record.upload_in_progress === 1)) {
       logger.info("Bailed download, upload in progress");
       continue;
     }
+
+    logger.info("download step 6-1");
 
     // Possible cases...
     // Case A: Perhaps the file was RENAMED on server. Delete from local
@@ -111,10 +111,10 @@ exports.recursiveDownload = async params => {
         }
       } // end Case A
 
-      // Case B: ...check last modified date and download of the file if newer (lets not download any folders)
+      // Case B: ...check last modified date and download the file if newer (lets not download any folders)
       // Convert the time to UTC and then get the unixtimestamp.
       else if (
-        _base.convertToUTC(node.modifiedAt) > _base.getFileLatestTime(record) &&
+        _base.convertToUTC(node.modifiedAt) > _base.getFileModifiedTime(record.file_path) &&
         record.file_name === path.basename(currentPath) &&
         record.is_file === 1
       ) {
@@ -173,6 +173,7 @@ exports.recursiveDownload = async params => {
     return;
   }
 };
+
 /**
  *
  * @param object params
@@ -185,53 +186,35 @@ exports.recursiveDownload = async params => {
 exports.recursiveUpload = async params => {
   const account = params.account;
   const watcher = params.watcher;
-  const rootNodeId = params.rootNodeId; // This should be the documentLibrary nodeId
+  let rootFolder = params.rootFolder;
 
   logger.info("upload step 1");
 
-  if (account.sync_enabled == 0 || account.sync_in_progress == 1) {
+  if (account.sync_enabled == 0 || account.upload_in_progress == 1) {
     logger.info("upload bailed");
     return;
   }
 
   logger.info("upload step 2");
-
-  // Get the folder path as /var/sync/documentLibrary or /var/sync/documentLibrary/watchedFolder
-  let rootFolder = path.join(
-    account.sync_path,
-    watcher.watch_folder.substring(
-      watcher.watch_folder.indexOf(`${watcher.site_name}/documentLibrary`)
-    ),
-    "**",
-    "*"
-  );
-
-  logger.info("upload step 3");
-
-  // This function will list all files/folders/sub-folders recursively.
-  let localFilePathList = glob.sync(rootFolder);
-
-  logger.info("upload step 4");
-
   // Following cases are possible...
   // Case A: File created or renamed on local, upload it
   // Case B: File modified on local, upload it
   // Case C: File deleted on server, delete on local
-  for (let filePath of localFilePathList) {
-    logger.info("upload step 5");
+  glob.sync(rootFolder).map(async filePath => {
+    logger.info("upload step 3");
     let localFileModifiedDate = _base.getFileModifiedTime(filePath);
 
     // Get the DB record of the filePath
     let record = await nodeModel.getOneByFilePath({
-      account: account,
+      account,
       filePath
     });
 
-    logger.info("upload step 6");
+    logger.info("upload step 4");
 
-    if (record && record.download_in_progress === 1) {
+    if (record && (record.download_in_progress == 1 || record.upload_in_progress == 1)) {
       logger.info("Bailed upload, download in progress. " + filePath);
-      continue;
+      return;
     }
 
     // Case A: File created or renamed on local, upload it
@@ -241,47 +224,37 @@ exports.recursiveUpload = async params => {
         account,
         watcher,
         filePath,
-        rootNodeId
+        rootNodeId: watcher.document_library_node
       });
-      continue;
     }
 
-    // Case B: File modified on local, upload it
-    else if (
-      record &&
-      localFileModifiedDate > record.last_uploaded_at &&
-      localFileModifiedDate > record.last_downloaded_at &&
-      Math.abs(localFileModifiedDate - record.last_downloaded_at) > 10 && // only upload if file wasnt downloaded in the last 10 seconds
-      record.is_file === 1
-    ) {
-      logger.info("File modified on local, uploading..." + filePath);
-      // Upload the local changes to the server.
-      await remote.upload({
-        account,
-        watcher,
-        filePath,
-        rootNodeId
-      });
-      continue;
-    }
+    logger.info("upload step 5");
 
-    // Case C: File deleted on server? delete on local
-    else if (
-      record &&
-      (record.last_uploaded_at > 0 || record.last_downloaded_at > 0)
-    ) {
-      logger.info("upload step 6-1");
+    // If the record exists in the DB (making sure upload is not in progress)
+    if (record) {
+      logger.info("upload step 6");
 
-      // Sleep for x seconds, so that server does not reject the request...
-      await _base.sleep(1000);
+      // Listen to the event
+      emitter.once('getNode' + record.node_id, async data => {
 
-      emitter.once('getNode' + record.node_id, data => {
+        // Case B: File modified on local, upload it
+        if (data.statusCode === 200 && data.record.is_file === 1 && localFileModifiedDate > _base.convertToUTC(data.response.entry.modifiedAt)) {
+          logger.info("File modified on local, uploading..." + filePath);
+          // Upload the local changes to the server.
+          await remote.upload({
+            account,
+            watcher,
+            filePath,
+            rootNodeId: watcher.document_library_node
+          });
+        }
 
-        // If the node is not found on the server, delete the file on local
-        if (data.statusCode === 404) {
+        // Case C: File deleted on server? delete on local
+        if (data && data.statusCode === 404 && data.record.download_in_progress == 0 && data.record.upload_in_progress == 0) {
           logger.info(
             "Node not available on server, deleting on local: " + data.record.file_path + " - " + data.record.id
           );
+          // If the node is not found on the server, delete the file on local
           rimraf(data.record.file_path, async () => {
             await nodeModel.forceDelete({
               account: data.account,
@@ -303,23 +276,33 @@ exports.recursiveUpload = async params => {
             });
           });
         }
-      });
+      }); // end event listener
 
+      // Make a request to the server to get the node details
       await remote.getNode({
         account,
         record
       });
-
-      logger.info("upload step 6-2. " + record.file_path);
-
+      logger.info("upload step 7");
     }
-    logger.info("upload step 7");
-  } // Filelist iteration end
-  logger.info("upload step 8");
-  localFilePathList = [];
-  logger.info("FINISHED UPLOAD...");
+
+    logger.info("upload step 8");
+
+    if (fs.statSync(filePath).isDirectory()) {
+      logger.info("upload step 9");
+      exports.recursiveUpload({
+        account,
+        watcher,
+        rootFolder: filePath + '/*'
+      });
+    }
+
+    logger.info("upload step 10");
+  }); // Filelist iteration end
+
+  logger.info("upload step 11");
   return;
-};
+}
 
 /**
  *
@@ -342,7 +325,9 @@ exports.deleteByPath = async params => {
 
   try {
     // Start the sync
-    await accountModel.syncStart(account.id);
+    await accountModel.syncStart({
+      account
+    });
 
     let records = await nodeModel.getAllByFileOrFolderPath({
       account: account,
@@ -357,12 +342,16 @@ exports.deleteByPath = async params => {
       });
     }
     // Set the sync completed time and also set issync flag to off
-    await accountModel.syncComplete(account.id);
+    await accountModel.syncComplete({
+      account
+    });
   } catch (error) {
     logger.error(`Error Message : ${JSON.stringify(error)}`);
     await errorLogModel.add(account.id, error);
     // Set the sync completed time and also set issync flag to off
-    await accountModel.syncComplete(account.id);
+    await accountModel.syncComplete({
+      account
+    });
   }
 };
